requirements.txt
Django>=3.0,<4.0
django_rest_framework
pandas
pymongo